// front-end/src/features/auth/model/permissions.ts
export enum PermissionEnum {
  CAN_READ_SCHEDULE = 'CAN_READ_SCHEDULE',

  CAN_CREATE_SESSIONS = 'CAN_CREATE_SESSIONS',
  CAN_READ_SESSIONS = 'CAN_READ_SESSIONS',
  CAN_UPDATE_OWN_SESSIONS = 'CAN_UPDATE_OWN_SESSIONS',
  CAN_UPDATE_ALL_SESSIONS = 'CAN_UPDATE_ALL_SESSIONS',

  CAN_CREATE_USERS = 'CAN_CREATE_USERS',
  CAN_READ_USERS = 'CAN_READ_USERS',
  CAN_UPDATE_USERS = 'CAN_UPDATE_USERS',

  CAN_CREATE_GROUPS = 'CAN_CREATE_GROUPS',
  CAN_READ_GROUPS = 'CAN_READ_GROUPS',
  CAN_UPDATE_GROUPS = 'CAN_UPDATE_GROUPS',

  CAN_CREATE_ROOMS = 'CAN_CREATE_ROOMS',
  CAN_READ_ROOM = 'CAN_READ_ROOM',
  CAN_READ_ROOMS = 'CAN_READ_ROOMS',
  CAN_UPDATE_ROOMS = 'CAN_UPDATE_ROOMS',

  CAN_READ_DEVICES = 'CAN_READ_DEVICES',
  CAN_UPDATE_DEVICES = 'CAN_UPDATE_DEVICES',

  CAN_READ_SETTINGS = 'CAN_READ_SETTINGS',
  CAN_UPDATE_SETTINGS = 'CAN_UPDATE_SETTINGS',
}